Introduce MapFuncStdLibSpecRunner
- Break StdLibSpecRunner into its own file
